move flat map serialization tests to kernel tests
